feat: Claim / Insurance / commande to run both server and client
-Finiche Vehicle Tabs Fix model delete

-Working on claim tab

Not commplite Implement delete vehicle functionality with confirmation modal and page refresh. Added ConfirmDeleteModal component to handle delete confirmations. Resolved warning related to DOM nesting and improved code structure. Added a new table in the database to handle the archive of deleted items. Also, added the model to add a new vehicle associated with the user that created it.

chore: Add Jest for testing UserRoutes for API testing